+++ fastplotlib/graphics/features/__init__.py
@@ -1,4 +1,4 @@
 from ._colors import ColorFeature, CmapFeature, ImageCmapFeature
 from ._data import PointsDataFeature, ImageDataFeature
 from ._present import PresentFeature
-from ._base import GraphicFeature
+from ._base import GraphicFeature, GraphicFeatureIndexable
